app/requirements.txt
Django==4.2.3
psycopg2-binary==2.9.6
djangorestframework>=3.0.0
djangorestframework-simplejwt==5.1.0